+++ go/enclave/rpc/rpc_utils.go
@@ -1,4 +1,4 @@
-package rpcencryptionmanager
+package rpc
 
 import (
 	"encoding/json"
